Rename handleEvent to handleInput
handleInput is more consistent because in a console, we are interested
about what the user gives us as input, not when or how. The word event
implies things like time or triggers. But in our case we are not
interested in such concepts, we just want to know what the user wants to
do next (by interpreting its input).