Fix a pointer binding bug in addAction
AbstractState::addAction was giving a wrong parameter to the method
that is associated with a menu entry. It was giving a reference to this
(casted to the subclasse type) instead of a pointer. But the first
parameter of a method is a pointer to the object, not a reference.
This led to a call of the method on a wrong object. This is now fixed.