Delay pop and clear from the the state stack
When a state calls pop, this state is deleted (its destructor gets
called) but when the pop is done, the stack returns to the caller of
the pop method, which is the state just popped out. Executing code
from a deleted object is undefined behavior, so by delaying the pop
(and the clear) we avoid this undefined behavior.