Do not delay StateStack::push
The push operation have not to be delayed because calling push do not
lead to undefined behavior. And by delaying it, substantial changes to
the state stack and to the client entry point have to be done, because
the first call to push (stack.push<HomeState>()) does not immediately
add the state to the stack, thus doPendingChanges need to be manually
called in order to continue the client properly.