Add CardsCollection class and configure CMake
The CardsCollection class has been implemented, and the CMakeLists.txt
files has been modified in order to compile the various compilation
units that are involved.

A file, common/Card.hpp, seems to be partially already written in the
branch cards. But the version of Card.hpp that is given here only
defines the type CardId, and thus these can easily be merged later.