Clean code in DecksManagementState::editDeck
The askForNumber method works now with std::size_t rather than int, this
is more consistent because the result of this methods is always used for
indexing an array.

editDeck has been rewritten in order to be more user-friendly (it asks
first for the replaced card, allowing to quit the editing menu without
have to specify the replacing card).