Temporaly disable return status in main
In the current state of the application, some work have to be done in
order to return a meaningful status from the main function, because the
code that checks for connexion error is no longer in main() but rather
in a function returning void. The best way to fix this problem is to
write custom exceptions classes, and throw instances of them, then catch
them in main().