Restructure build process, making it more modular
Almost all sources files are now linked in temporary libraries rather
than directly in the executable. Only the entry point (the file with a
main() function) is direclty linked to the executable.

This allows the unit test executable to reuse and link to these
libraries rather than compile once again the needed source files.