Clear up : there are no spell constraints, just creature and player constraints
A constraint is a rule that is checked at some point throughout the game and that has a time dimension (a number of turns). They can be considered like passive or turn-based effects. Pretty much any effect that isn't called immediately or at the end of a creature's life will be stored as a constraint.