Add COLLATE NOCASE to Cards.name
We decide to disallow user to sort cards freely. So COLLATE NOCASE will
allow us to sort by name.

I do not have internet so I have not english dictionnary to know the
order of words in sentences and to find the linnk of the issue
("question") about this on github. Sorry.