Make cards declaration external to CardData.hpp
Variables (arrays for instance) cannot be declared in a header. Reason is
that if the header is included by several files in the same compiled unit,
there exists several variables with the same name and then the compiler
doesn't understand anything.

To avoid this, the solution is to declare the variables somewhere else (in
a source *.cpp file) and possibly, have the declaration `extern <type> <name>;`
in a header so that every file can be compiled even though it does not have
**yet** the variable to access.