Implements decks management in Client
This commit also set a default value for the name parameter in the
constructor of the class Deck. Without it, this is impossible to
construct a `std::vector<Deck>` because std::vector requires that its
elements are default-constructibles. This is not the case if the string
argument of the Deck constructor has no default value.