Merge class Board with class GameThread
The two classes (Board and GameThread) have been developed at two different times
but with the same general purpose: being the responsible for one whole game,
from the early initializations to the very end when ending signals are sent.

As discussed with @TheoVerhelst and @StanIsAdmin on Friday, these two classes
should be merged as one. This is the purpose of this commit.

The class Board doesn't exist anymore then and is totally replacedby GameThread.