Use std::lock_guard rather than manually lock
This is more exception-safe because we can't know if an exception will
be thrown or not in a method, so if the unlock call is at the end of the
method and the method throw before the unlock, the mutex will never be
released, and some thread will starve.

This is sad and bad, so the solution is to use scoped lock:
std::lock_guard is an object that lock a given mutex at construction and
unlock it at deletion. So if the method throw, the lock_guard is deleted
by the stack unwinder, and the mutex is unlocked, and everyone is happy.