Fix server deadlock and implement game quitting
Server deadlock was due to the automatic lock of std::unique_lock performed
inside the constructor. So when calling std::unique_lock::lock right after,
a unique lock was trying to be locked again which is a deadlock.

Game is implemented on the following principle: a GameState object is created
and the function to have it working is called **inside** the constructor.
This way when the object is eventually created, the game is over and then
the object is pseudo-destroyed by calling stackPop (from the lobby manager
which pushes the GameState).