Finish to move normal socket in the Player class
This implies some changes in the board changes managing: now the changes
that occur on the board (such as health or hand card changes) are logged
in a private attribute (a sf::Packet) and the class GameThread sometimes
asks the Player if there are some changes to send to the client on the
special socket. If so, the changes are sent.

And everything that is related to the normal socket is now managed into
the Player class, such as receiving the choosend deck.