Remove TransferType::END_OF_TRANSMISSION
A sf::Packet knows when it has no more data to give to the program which
read it, and this is precisely the goal of the method
sf::Packet::endOfTranmission, so there is no need to have an header that
tell when there is no more data.

This removal also implies some changes in the method which read a packet
in the client side (GameState::handlePacket).