Improve end of game + give new card to player (database match still TODO)
The winner now sees the card he received at the end of the game. But it is
not added to the database (maybe let's have @areynoua or @TheoVerhelst on this?)
yet. There are two new values in TransferType: WINNER and LOSER which are sent
**after** GAME_OVER in the end-of-game-transmission.

Reason why keeping GAME_OVER is in the code.