Make CardToSelect a scoped enum
Scoped enumerations are preferable over unscoped ones if they aren't
intented to be used in arithmetics:

* a scoped enumeration (`enum class` or `enum struct`) is used to have
a type with distincts values, no matter of how it is encoded. In
particular, we don't care about a relational order between the values,
we just want to know if they are equal or not;

* an unscoped enumeration (`enum`) is used to have a type with
distincts values, and to be able to do arithmetics operations on it
(such as bitwise masking, addition, etc). This case is far less common.

Although the underlying type of a scoped enum can be specified, they
can't be implicitely converted to it, enforcing the fact that scoped
enums are not meant to be used in arithmetic expressions.

Obviously, CardToSelect belongs to the first case.