Send more informations to client at end of the game
The struct EndGame gives informations about why and how the game ended,
and an instance is sent to each player at game end. A result of this
change is that TransferType::WINNER and LOSER are no longer needed,
because this information is already in EndGame.

The receiving data are also properly handled in client side.