Use parenthesis rather than braces for init vector
Using constructor with braces on a `std::vector<sf::Uint32>` with an
unsigned type leads, on some machines (e.g. mine), to use the
constructor that takes a std::initializer_list as argument, making the
compilation fail because the types mismatch (long unsigned vs unsigned
int).

By using parentheses instead, the constructor that takes a size and that
default-construct the given number of elements is called instead, wich
is the desired behavior in this case.