Add UnableToConnectException to distinguish errors when connecting
Because all connection and authentication exceptions were
std::runtime_error, there was no way to know if the connection failed
because of a socket error or something else (e.g. invalid user name),
and this was annoying in the code that tests 10 ports, because the same
message was printed for every error.