Don't forget to increment it in Constraints::getSumTimesValues
It seems that the iterator is not incremented in the loop that browses
the list of constrainsts when the iterator points to a elements to take
in account (i.e. to add to the sum). @StanIsAdmin can you confirm that
this was not the desired behavior?

This commit also format the code by adding and removing spaces in some
places, I've a bit of time to loose.