Use unique_ptr instead of raw pointers for Card in Player
This is for the same reason as in other places of the project, a smart
pointer is more explicit about the ownership of the pointed object, and
they are always deleted automatically.

This involves many changes, because in every place where a simple copy
of the pointer were sufficient, now we have to call std::move to
explicitely move the ownership of the resource.

Now the Player class need the Creature class to be a complete type, so
we have to include Creature.hpp in Player.hpp, and make a foward
declaration of Player in Creature.hpp.

I also reformatted a bit the code to be more consistent with the rest of
the project.