Fix #105
Don't forget to remove 1 at the index because they start with value 1 (SQLite).
Also, Creature::_effectMethods has changed from a `std::function<>[]` to
a `std::array<std::function, X>`. This allows to use the `at` method
which throws a `std::out_of_range` in case of wrong index. This can be somehow helpful!