Finish the implementation of the new state design
This is a complicated design, with a diamond inheriting. Here I explain
a bit how it works:
* We have a base class, AbstractState;
* Then we have TerminalAbstractState and GuiAbstractState that inherits
from AbstractState;
* And we have an abstract class for each state, such as
AbstractHomeState;
* Each state has also two concrete classes: TerminalHomeState that
* inherits from TerminalAbstractState and AbstractHomeState, and the
* corresponding class for the GUI;

This design, although complicated, allows to separate very efficiently
the terminal/gui logic, the logic specific to a menu, and the logic for
every state.

**Note:** the classes for the GUI aren't yet written, but their
declarations will be similar to the Terminal* ones.