Replace clear system call in main client function
By writing it into TerminalAbstractState::display, the mechanism of
redefining display in child classes become useless, because the first
instruction of TerminalAbstractState::display would be to clear the
screen, thus all informations printed by the child class is lost.

So the clear call has to be before the call to the child class' display
method.