Add the context structure
To see the description and the use of this struct, see its documentation
in front of its declaration.

Now each state takes a reference to the Context rather than two
references for the client and the state stack. This allows to add other
objects in the context without have to rewrite every constructor.