Add more logs when entering turns
The effects a monster could make when beginning a turn are:
+ change its health/attack/shield;
+ change the player's health;
+ same on the other side of the board.

So these data must be logged to be sent to the players.

Also, maybe there can be optimisations there eventually because all
of these are not modified everytime but this is definitely not a big deal
and is totally not to do now.