Fix #110
The client expected the server to send a vector of additional inputs when
a card is set on board and the server sent it if and only if the board was
not empty. And when it was, nothing was sent, causing the client to be garbaged.

Now, this communication begins with a header (TransferType) equal to ACKNOWLEDGE
or FAILURE depending on whether or not the card was set on board.