Use only const references for effects
Define a struct containing a const reference to effect parameters along with an index indicating what value should be read next. When reading a value, index must be incremented.
Expected behavior is : when a function that requires effect parameters is called, the struct is passed as an argument, the index is copied but the vector of parameters is not, only its reference is.