Implement GuiFriendsManagementState::removeFriend()
To achieve this, I've implemented a new class: `InputBox`.
Principle is to create a "pop-up" like a message box but with an
Editbox on it to get a user input. The way it works will be detailed
in next commit (but first, let's get some sleep!); just: a callback
must be given to the object so it knows what to do when "Ok" is pressed
and to cancel the current operation, destroying the widget by clicking
the upper-right cross is sufficient.