Fix widget allocation issue
This is done by adding two virtual methods to AbstractState, onPush and
onPop. These methods are called when a stack is pushed or poped, and
allow the state to properly show/hide their widgets, to let other state
on top of them to only show their owns.

Although this commit effectively fix an issue, it doesn't fix #121 as
expected.