Fix #121
The bug was caused by two things:
1. The static attributes for containing widgets are very, very bad idea.
That lead to the same widget being bind multiples times when the user go
back to a menu already visited, so the related action called multiple
times...

2. The management of the state by the StateStack also may caused subtle
bugs because the states where not properly destructed once popped. I
don't exactly know in which way it was wrong, but I know that the system
I written back is safe and simpler to read, although maybe less
efficient. But we don't care about efficiency, if it take 4 days to
debug it then!