Factorize onPush, onPop and destructor of Gui states
All these methods did the same thing: hide, show or removing widgets.
So this is possible to factorize this behavior as well, this is done in
the GuiAbstractState class: the derived state register to its
GuiAbstractState base which widgets need to be hidden/shown/removed, and
the base do it itself when needed.