Make CardGui inheriting from sf::Transformable
I didn't do that at first because I thought that we will only have to
move the cards arounds, but not to scale them nor rotate them. But it
may be really useful to scale cards in the menu, so that we can display
several of them at once in the screen (with the actual size, we owuld be
able to display maximum around 10 cards).

And rotating will be useful when making the graphics of the game board.