Fix displayed deck indexes, lots of TODOs
Cards are not unlocked in some order, card's ids should not be used for choosing a card within a deck or for anything else than requesting card data to a database, or identifying a card. Therefore, the way decks are handled should probably be changed or deserves some comments, cause it's confusing. Anyway, at least cards are DISPLAYED now.