Change the command line option --gui to --no-gui
In #91 we stated that the command line options for this programm will be
--no-gui for running the terminal version, and no option for running the
graphical version. I used instead --gui at first because we were using
more often the terminal version, but since the GUI is well deployed,
this would be better to use --no-gui as stated at first.

This commit also allocates the Application object on the stack rather
than on the heap, this is shorter and faster (no virtual call).