Re-fix deck indices in TerminalDecksManagementState
When the user want to modify a deck, the client asked him to select a
card from the deck, or 0 to exit, but since the indices started at 0,
there was no way to select the first card.

Also, that was the content of the card collection that was displayed
rather than the deck.