Use new method of AbstractDecksManagementState in TerminalDecksManagementState
And when we ask the user to choose a card from its collection when
editing a deck, the displayed message before displaying available cards
is changed to be more consistent. This is because we don't show anymore
the full collection, but only a part of it that is relevant according to
the modified deck.