Fix #163
The operator >>(Packet&, Deck&) creates a new deck. Then the assertion failed
when setting the 2nd value to itself (actually) which shouldn't be a problem.
Now, one more parameter is available to decide whether the whole deck must
be searched for > 2 copies (when changing a card) or only into the current index
(when creating a deck card by card).